move set_definition method to container
